refactor: MapBoundsDisplay 위치 이동
