+++ src/components/recruitment/develop/MapBoundsDisplay.module.scss
